pyproject.toml
[project]
name = "call-my-agent"
version = "0.1.0"
description = "Writer + Editor loop using Semantic Kernel with Azure OpenAI"
dependencies = [
    "python-dotenv",
    "semantic-kernel"
]

[project.scripts]
call-my-agent = "main:main"

[tool.uv]
python = ">=3.13"
